Implementing a RoundCounter in the MenuBar
An EventBased RoundCounter has been set in
the AbstractPresenterWithChatWithGame to display
the Current Round the Game is in.
Its set in the onNextPlayerMessage to add one Round
to the Counter as soon as this Message is posted onto
the EventBus.
Its visibilities have been set to display the Counter
with the onStartSessionButtonPressed and undisplay
and reset it with the onReturnToPreGameLobbyMessage
found in the AbstractPresenterWithChatWithGameWithPreGamePhase.

Author: Aldin Dervisi <aldin.dervisi@uni-oldenburg.de>