Setting the RoundCounter on different base
Instead of increasing the Round by 1 when a Turn has ended,
the RoundCounter will be increased by 1 as soon as all players
in a specific Lobby have had a turn and ended it.
In the Game.java there has been implemented a getFirst method
which receives the first player to make a turn in a Lobby.
With that, every time the first player is on turn again,
the RoundCounter will be increased by one.

Author: Aldin Dervisi <aldin.dervisi@uni-oldenburg.de>